requirements.txt
fastf1-livetiming @ git+https://github.com/br-g/fastf1-livetiming.git
psycopg2-binary==2.9.6
loguru==0.7.0
python-dotenv==1.0.0
asyncpg==0.27.0